requirements.txt
rumps
screeninfo
mss
Pillow
numpy
